requirements.txt
pillow
Pyqt6=6.6.1